Revert "Create a Simple Cache Strategy To Allow Fallback"
This reverts commit 5240d427b8bf6501603945aecd8b20d82650feb4.
